requirements.txt
python-telegram-bot==13.15
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.7